requirements.txt
django
djangorestframework
djangorestframework-simplejwt
python-dotenv
psycopg2-binary
gunicorn